+++ udiskie/common.py
@@ -14,6 +14,7 @@ __all__ = [
     'sameuuid',
     'setdefault',
     'extend',
+    'cachedproperty',
     'decode_ay',
     'exc_message',
     'format_exc',
@@ -32,7 +33,7 @@ class Emitter:
 
     def __init__(self, event_names=(), *args, **kwargs):
         """Initialize with empty lists of event handlers."""
-        super(Emitter, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self._event_handlers = {}
         for evt in event_names:
             self._event_handlers[evt] = []
@@ -75,6 +76,20 @@ def extend(a: dict, b: dict) -> dict:
     res = a.copy()
     res.update(b)
     return res
+
+
+def cachedproperty(func):
+    """A memoize decorator for class properties."""
+    key = '_' + func.__name__
+    @wraps(func)
+    def get(self):
+        try:
+            return getattr(self, key)
+        except AttributeError:
+            val = func(self)
+            setattr(self, key, val)
+            return val
+    return property(get)
 
 
 # ----------------------------------------